// frontend/src/lib/api/models/LecturerRosterRiskRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LecturerRosterRiskRow } from './LecturerRosterRiskRow';
export type LecturerRosterRiskRead = {
    moduleId: string;
    moduleTitle: string;
    needsSupportCount: number;
    rows: Array<LecturerRosterRiskRow>;
};