// frontend/src/lib/api/models/RiskReasonRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RiskReasonRead = {
    code: string;
    severity: string;
    metricKeys: Array<string>;
    lecturerText: string;
    studentText: string;
    supportingMetrics: Record<string, (string | number | null)>;
};